Firing Platform Animation

+++ BoardBall_Game/Ball.cpp
@@ -394,7 +394,7 @@ void ABall::Draw_Parachute(HDC hdc, RECT &paint_area)
 {
 	RECT intersection_rect;
 	const int scale = AsConfig::Global_Scale;
-	const int dome_size = Parachute_Size * scale / 2.0;
+	const int dome_size = int(Parachute_Size * scale / 2.0);
 	int ball_center_x, ball_center_y;
 
 	if (! IntersectRect(&intersection_rect, &paint_area, &Parachute_Rect) )
@@ -414,7 +414,7 @@ void ABall::Draw_Parachute(HDC hdc, RECT &paint_area)
 
 	Ellipse(hdc, Parachute_Rect.left + 11 * scale, Parachute_Rect.top + 5 * scale + 1, Parachute_Rect.left + 14 * scale + 1, Parachute_Rect.top + 9 * scale - 1);
 
-	ball_center_x = Parachute_Rect.left + Parachute_Size * scale / 2.0;
+	ball_center_x = int(Parachute_Rect.left + Parachute_Size * scale / 2.0);
 	ball_center_y = Parachute_Rect.bottom;
 
 	AsConfig::White_Color.Select(hdc);
